feat: 프론트 페이지 헤더 메인 푸터 레이아웃 일괄적용 중...
- 인덱스 페이지 관련 css,html 수정
- index 컨트롤러 토큰에서 role 검사 로그인 검사 aop로 모든 컨트롤러에 적용
- 추후 다른 페이지에도 적용 예정
