+++ pom.xml
@@ -77,6 +77,23 @@
             <scope>provided</scope>
 
         </dependency>
+        <!-- aop 의존성 추가 -->
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjrt</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjweaver</artifactId>
+            <scope>compile</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjtools</artifactId>
+        </dependency>
+
 <!--레디스 관련 pom 추가-->
         <dependency>
             <groupId>org.springframework.boot</groupId>
