+++ src/main/java/com/nhnacademy/bookstorefront/main/controller/IndexController.java
@@ -1,13 +1,8 @@
 package com.nhnacademy.bookstorefront.main.controller;
 
-import com.nhnacademy.bookstorefront.main.client.AuthenticationClient;
 import com.nhnacademy.bookstorefront.main.client.BookClient;
-import com.nhnacademy.bookstorefront.main.dto.BookDetailResponseDto;
-import com.nhnacademy.bookstorefront.main.dto.PagedResponse;
 import com.nhnacademy.bookstorefront.main.dto.book.SellingBookResponseDto;
-import com.nhnacademy.bookstorefront.main.service.AuthenticationService;
 import com.nhnacademy.bookstorefront.main.service.CookieService;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
@@ -25,8 +20,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 @RequiredArgsConstructor
 public class IndexController {
     private final BookClient bookClient;
-    private final AuthenticationService authenticationService;
-    private final AuthenticationClient authenticationClient;
     private final CookieService cookieService;
 
     @GetMapping({"/index", "/"})
@@ -45,16 +38,6 @@ public class IndexController {
         // Feign 클라이언트를 통해 페이징된 데이터 요청
         Page<SellingBookResponseDto> response = bookClient.getBooks(page, size, sortBy, sortDir).getBody();
 
-        boolean isLoggedIn = authenticationService.isLoggedIn(request);
-        String role = null;
-
-        if(isLoggedIn) {
-            String token = getTokenFromCookies(request);
-            if(token != null) {
-                role = authenticationClient.getRoleFromToken("Bearer " + token).getBody();
-            }
-        }
-
 
         model.addAttribute("books", response.getContent());
         model.addAttribute("currentPage", response.getNumber());
@@ -62,21 +45,8 @@ public class IndexController {
         model.addAttribute("sortBy", sortBy);
         model.addAttribute("sortDir", sortDir);
         model.addAttribute("pageSize", size);
-        model.addAttribute("isLoggedIn", isLoggedIn);
-        model.addAttribute("role", role);
 
         return "index"; // index.html 렌더링
-    }
-
-    private String getTokenFromCookies(HttpServletRequest request) {
-        if (request.getCookies() != null) {
-            for (Cookie cookie : request.getCookies()) {
-                if ("accessToken".equals(cookie.getName())) {
-                    return cookie.getValue();
-                }
-            }
-        }
-        return null;
     }
 
 }
