+++ src/main/resources/static/css/layout/layout.css
@@ -1,8 +1,14 @@
-body {
+html, body {
     font-family: 'Arial', sans-serif;
     background-color: #f9f9f9;
     margin: 0;
     padding: 0;
+    height: 100%;
+}
+
+body {
+    display: flex;
+    flex-direction: column;
 }
 
 header {
@@ -35,11 +41,127 @@ header .icons a:hover {
     color: #f1f1f1;
 }
 
+main {
+    display: flex;
+    gap: 20px;
+    flex-direction: column;
+    padding: 20px;
+    flex: 1;
+}
+
+/* 검색바와 정렬바 */
+.search-bar, .sort-bar {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    margin-bottom: 20px; /* 요소 간의 간격 */
+}
+
+
+.sidebar {
+    width: 220px;
+    background-color: #fff;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    padding: 20px;
+}
+
+.sidebar .group {
+    display: block; /* a 태그를 블록 요소로 변경 */
+    text-decoration: none; /* 기본 밑줄 제거 */
+    color: inherit; /* 기본 색상 상속 */
+    padding: 12px;
+    border-radius: 4px;
+    cursor: pointer;
+    margin-bottom: 10px;
+    font-size: 16px;
+    transition: background-color 0.3s;
+    text-align: center;
+}
+
+.sidebar .group.active,
+.sidebar .group:hover {
+    background-color: #ddd;
+    color: black;
+}
+
+.content {
+    flex: 1;
+    padding: 20px;
+    background-color: #fff;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.content h2 {
+    font-size: 24px;
+    color: #333;
+    margin-bottom: 20px;
+}
+
+.section {
+    display: none;
+}
+
+.section.active {
+    display: block;
+}
+
+
+.table-wrapper {
+    overflow-x: auto;
+    margin-bottom: 20px;
+}
+
+table {
+    width: 100%;
+    border-collapse: collapse;
+    margin-top: 10px;
+}
+
+th, td {
+    padding: 10px;
+    text-align: left;
+    border: 1px solid #ddd;
+}
+
+th {
+    background-color: #f1f1f1;
+}
+
+td {
+    background-color: #fff;
+}
+
+.alert {
+    padding: 15px;
+    margin-bottom: 20px;
+    border-radius: 5px;
+    font-size: 16px;
+}
+
+.alert.success {
+    background-color: #d4edda;
+    color: #155724;
+    font-weight: bold;
+    border: 1px solid #c3e6cb;
+}
+
+.alert.error {
+    background-color: #f8d7da;
+    color: #721c24;
+    font-weight: bold;
+    border: 1px solid #f5c6cb;
+}
+
+
 footer {
     background-color: #333;
     color: white;
     padding: 15px 20px;
     text-align: center;
+    left: 0;
+    right: 0;
 }
 
 footer a {
@@ -51,3 +173,42 @@ footer a {
 footer a:hover {
     text-decoration: underline;
 }
+
+.pagination {
+    display: flex;
+    justify-content: center;
+    padding: 20px 0;
+    list-style: none;
+}
+
+/*.page-item {*/
+/*    margin: 0 5px;*/
+/*}*/
+
+/*.page-item a {*/
+/*    display: block;*/
+/*    padding: 10px 15px;*/
+/*    color: #333;*/
+/*    background-color: #fff;*/
+/*    border: 1px solid #ddd;*/
+/*    border-radius: 4px;*/
+/*    text-decoration: none;*/
+/*    transition: background-color 0.3s, color 0.3s;*/
+/*}*/
+
+/*.page-item a:hover {*/
+/*    background-color: #f1f1f1;*/
+/*    color: #000;*/
+/*}*/
+
+/*.page-item.active a {*/
+/*    background-color: #333;*/
+/*    color: #fff;*/
+/*    border-color: #333;*/
+/*}*/
+
+/*.page-item.disabled a {*/
+/*    color: #ccc;*/
+/*    pointer-events: none;*/
+/*    background-color: #f9f9f9;*/
+/*}*/
