+++ src/main/resources/static/index.css
@@ -1,82 +1,100 @@
-/* 기본 스타일 */
-body {
-    font-family: 'Arial', sans-serif;
-    background-color: #f9f9f9;
-    margin: 0;
-    padding: 0;
-}
-
-/* 헤더 */
-header {
-    background-color: #333;
-    color: white;
-    text-align: center;
-    padding: 20px 0;
-    font-size: 24px;
-    font-weight: bold;
-}
-
-/* 검색 바 */
-.search-bar {
-    text-align: center;
-    margin: 20px 0;
+main {
+    display: flex;
+    flex-direction: column;
+    padding: 20px;
+    flex: 1;
 }
 
 .search-bar input {
+    flex: 1;
     padding: 10px;
-    width: 300px;
     font-size: 16px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
 }
 
 .search-bar button {
+    flex: auto;
     padding: 10px 15px;
     background-color: #333;
     color: white;
     border: none;
+    border-radius: 4px;
     cursor: pointer;
 }
 
-/* 책 목록 */
+.search-bar button:hover {
+    background-color: #444;
+}
+
+.sort-bar select {
+    flex: auto;
+    padding: 10px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+}
+
+/* 책 목록 스타일 */
 .book-list {
     display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+    grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
     gap: 20px;
-    padding: 20px;
 }
 
 .book {
     text-align: center;
-    background-color: white;
-    padding: 15px;
-    border-radius: 10px;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+    background-color: #fff;
+    padding: 10px;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 }
 
-/* 페이징 버튼 */
+.book img {
+    width: 100%;
+    height: auto;
+    border-radius: 4px;
+    margin-bottom: 10px;
+}
+
+.book div {
+    font-size: 14px;
+    color: #333;
+}
 .pagination {
     display: flex;
     justify-content: center;
-    margin: 20px 0;
+    padding: 20px 0;
     list-style: none;
-    gap: 10px;
-    padding: 0;
 }
 
 .page-item {
+    margin: 0 5px;
+}
+
+.page-item a {
+    display: block;
     padding: 10px 15px;
-    background-color: #f9f9f9;
+    color: #333;
+    background-color: #fff;
     border: 1px solid #ddd;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
+    border-radius: 4px;
+    text-decoration: none;
+    transition: background-color 0.3s, color 0.3s;
 }
 
-.page-item.active {
+.page-item a:hover {
+    background-color: #f1f1f1;
+    color: #000;
+}
+
+.page-item.active a {
     background-color: #333;
-    color: white;
+    color: #fff;
+    border-color: #333;
 }
 
-.page-item.disabled {
-    color: #aaa;
+.page-item.disabled a {
+    color: #ccc;
     pointer-events: none;
+    background-color: #f9f9f9;
 }
