+++ src/main/resources/templates/cart/myCart.html
@@ -1,50 +1,13 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html xmlns:th="http://www.thymeleaf.org" lang="ko">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
     <link rel="stylesheet" href="/detail.css?v=1.1">
+    <link rel="stylesheet" href="/css/layout/layout.css">
     <title>장바구니</title>
-
     <style>
-        /* 공통 스타일 */
-        html, body {
-            font-family: 'Arial', sans-serif;
-            background-color: #f9f9f9;
-            margin: 0;
-            padding: 0;
-        }
-
-        header {
-            background-color: #333;
-            padding: 15px 20px;
-            color: white;
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-        }
-
-        header .logo {
-            text-decoration: none;
-            color: inherit;
-            font-size: 24px;
-            font-weight: bold;
-        }
-
-        header .menu a,
-        header .icons a {
-            color: white;
-            text-decoration: none;
-            margin: 0 15px;
-            font-size: 16px;
-        }
-
-        header .menu a:hover,
-        header .icons a:hover {
-            color: #f1f1f1;
-        }
-
         main {
             padding: 20px;
             position: relative; /* 상대 위치 지정 */
@@ -129,20 +92,7 @@
     </style>
 </head>
 <body>
-<header>
-    <a href="/" class="logo">📚 책사조잉 📚</a>
-    <nav class="menu">
-        <a href="/">Home</a>
-    </nav>
-    <div class="icons">
-        <a th:if="${role == 'ADMIN'}" href="/adminpage" id="adminpage-btn">AdminPage</a>
-        <a th:if="${role == 'USER'}" href="/mypage" id="mypage-btn">MyPage</a>
-        <a th:if="${role == null and !isLoggedIn}" href="/non-member/order" id="cart-btn">비회원 주문</a>"
-        <a th:if="${role == null and !isLoggedIn}" href="/login" id="login-btn">로그인</a>
-        <a th:if="${role == null and !isLoggedIn}" href="/signup" id="signup-btn">회원가입</a>
-        <a th:if="${isLoggedIn}" href="/logout" id="logout-btn">로그아웃</a>
-    </div>
-</header>
+<th:block th:replace="~{layout/layout :: headerFragment}"></th:block>
 <main class="container mt-5">
     <div id="cartItems">
         <div th:each="item : ${cartItems}" class="cart-item row align-items-center">
