+++ src/main/resources/templates/index.html
@@ -1,173 +1,15 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html xmlns:th="http://www.thymeleaf.org" lang="ko">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/index.css?v=1.0">
+    <link rel="stylesheet" href="/index.css">
+    <link rel="stylesheet" href="/css/layout/layout.css">
     <title>📚 책사조잉 📚</title>
-    <style>
-        /* 공통 스타일 */
-        body {
-            font-family: 'Arial', sans-serif;
-            background-color: #f9f9f9;
-            margin: 0;
-            padding: 0;
-        }
 
-        header {
-            background-color: #333;
-            padding: 15px 20px;
-            color: white;
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-        }
-
-        header .logo {
-            text-decoration: none;
-            color: inherit;
-            font-size: 24px;
-            font-weight: bold;
-        }
-
-
-        header .menu a,
-        header .icons a {
-            color: white;
-            text-decoration: none;
-            margin: 0 15px;
-            font-size: 16px;
-        }
-
-        header .menu a:hover,
-        header .icons a:hover {
-            color: #f1f1f1;
-        }
-
-        main {
-            padding: 20px;
-        }
-
-        /* 검색바와 정렬바 */
-        .search-bar, .sort-bar {
-            display: flex;
-            align-items: center;
-            gap: 10px;
-            margin-bottom: 20px;
-        }
-
-        .search-bar input {
-            flex: 1;
-            padding: 10px;
-            font-size: 16px;
-            border: 1px solid #ddd;
-            border-radius: 4px;
-        }
-
-        .search-bar button {
-            padding: 10px 15px;
-            background-color: #333;
-            color: white;
-            border: none;
-            border-radius: 4px;
-            cursor: pointer;
-        }
-
-        .search-bar button:hover {
-            background-color: #444;
-        }
-
-        .sort-bar select {
-            padding: 10px;
-            border: 1px solid #ddd;
-            border-radius: 4px;
-        }
-
-        /* 책 목록 스타일 */
-        .book-list {
-            display: grid;
-            grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
-            gap: 20px;
-        }
-
-        .book {
-            text-align: center;
-            background-color: #fff;
-            padding: 10px;
-            border-radius: 8px;
-            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-        }
-
-        .book img {
-            width: 100%;
-            height: auto;
-            border-radius: 4px;
-            margin-bottom: 10px;
-        }
-
-        .book div {
-            font-size: 14px;
-            color: #333;
-        }
-
-        /* 버튼 */
-        button {
-            padding: 12px 20px;
-            background-color: #333;
-            color: white;
-            font-size: 16px;
-            border: none;
-            border-radius: 8px;
-            cursor: pointer;
-            transition: background-color 0.3s;
-        }
-
-        button:hover {
-            background-color: #444;
-        }
-
-        footer {
-            background-color: #333;
-            color: white;
-            padding: 15px 20px;
-            text-align: center;
-        }
-
-        footer a {
-            color: #00aaff;
-            text-decoration: none;
-            margin: 0 5px;
-        }
-
-        footer a:hover {
-            text-decoration: underline;
-        }
-    </style>
 </head>
 <body>
-<header>
-    <a href="/" class="logo">📚 책사조잉 📚</a>
-    <nav class="menu">
-        <a href="/">Home</a>
-        <a href="#">국내도서</a>
-        <a href="#">외국도서</a>
-        <a href="#">추천도서</a>
-    </nav>
-    <div class="icons">
-        <!-- 로그인 상태에 따라 버튼 표시 -->
-        <a th:if="${role == 'ADMIN'}" href="/adminpage" id="adminpage-btn">AdminPage</a>
-        <a href="/non-member/order">비회원 주문조회</a>
-        <a th:if="${role == 'USER'}" href="/mypage" id="mypage-btn">MyPage</a>
-        <a th:if="${role == null and !isLoggedIn}" href="/guests/carts">비회원 장바구니</a>
-        <a th:if="${role == 'USER' and isLoggedIn}" href="/carts" id="cart-btn">장바구니</a>
-        <a th:if="${role == null and !isLoggedIn}" href="/login" id="login-btn">로그인</a>
-        <a th:if="${role == null and !isLoggedIn}" href="/signup" id="signup-btn">회원가입</a>
-        <a th:if="${isLoggedIn}" href="/logout" id="logout-btn">로그아웃</a>
-
-    </div>
-</header>
-
-
+<th:block th:replace="~{layout/layout :: headerFragment}"></th:block>
 <main>
     <div class="search-bar">
         <form id="searchForm" method="GET" action="/search/books">
@@ -204,37 +46,30 @@
         </div>
     </div>
 
-</main>
-
 
 <nav aria-label="Page navigation">
     <ul class="pagination">
-        <!-- 이전 버튼 -->
+        <!-- Previous button -->
         <li class="page-item" th:classappend="${pageNumber == 0} ? 'disabled'">
             <a th:href="@{/index(page=${pageNumber != null && pageNumber > 0 ? pageNumber - 1 : 0}, size=${pageSize != null ? pageSize : 16}, sortBy=${sortBy}, sortDir=${sortDir})}">이전</a>
         </li>
 
-        <!-- 페이지 번호 버튼 -->
+        <!-- Page number buttons -->
         <li class="page-item" th:each="page, stat : ${#numbers.sequence(0, totalPages != null ? totalPages - 1 : 0)}"
             th:classappend="${page == (pageNumber != null ? pageNumber : 0)} ? 'active'">
             <a th:href="@{/index(page=${page}, size=${pageSize != null ? pageSize : 12}, sortBy=${sortBy}, sortDir=${sortDir})}"
                th:text="${page + 1}"></a>
         </li>
 
-        <!-- 다음 버튼 -->
+        <!-- Next button -->
         <li class="page-item"
             th:classappend="${pageNumber != null && pageNumber == (totalPages != null ? totalPages - 1 : 0)} ? 'disabled'">
             <a th:href="@{/index(page=${pageNumber != null && pageNumber < (totalPages != null ? totalPages - 1 : 0) ? pageNumber + 1 : (totalPages != null ? totalPages - 1 : 0)}, size=${pageSize != null ? pageSize : 16}, sortBy=${sortBy}, sortDir=${sortDir})}">다음</a>
         </li>
     </ul>
 </nav>
-
-<footer>
-    <p>&copy; 2025 PlzBuyBook. 모든 권리 보유.</p>
-    <p><a href="https://github.com/nhnacademy-be8-plzbuybook/bookstore-shoppingmall" target="_blank"> PlzBuyBook
-        GitHub</a></p>
-    <p><a href="/privacy">개인정보처리방침</a> | <a href="/terms">이용약관</a></p>
-</footer>
+</main>
+<th:block th:replace="~{layout/layout :: footerLayout}"></th:block>
 <script src="/script/books.js"></script>
 
 <script>
